pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "yarbo"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    # Pinned to commit SHA — pre-1.0, ships from main, bumps via PR + smoke test
    "python-yarbo @ git+https://github.com/markus-lassfolk/python-yarbo@72d0262b50f7b8bbabcbe3fe2dd06648260dcb4d",
    "paho-mqtt>=2.1",
    "playwright>=1.44",
    "requests>=2.32",
]

[project.scripts]
yarbo = "yarbod.cli:main"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/yarbod"]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "hypothesis>=6.100",
    "requests-mock>=1.12",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=yarbod --cov-report=term-missing --cov-fail-under=80"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/yarbod"]
omit = ["tests/*"]
